Update integers to unsigned and add new .bat file
This major commit updates a substantial amount of the code base to use unsigned integers (uint32_t) instead of integers, for improved efficiency and to align with data type requirements. Additionally, the commit introduces a new Windows batch file (`web_dm.bat`) in the tools directory to copy PHP files across directories. A range of files across the project are affected, with key changes in `SemanticJob.cpp`, `Tokenizer.cpp` and `SCBE_X64.cpp`. Several tests and scripts have also been updated.

+++ src/Naming.cpp
@@ -454,7 +454,7 @@ Utf8 Naming::aKindName(ScopeKind kind)
     return article;
 }
 
-Utf8 Naming::niceArgumentRank(size_t idx)
+Utf8 Naming::niceArgumentRank(uint32_t idx)
 {
     switch (idx)
     {
@@ -469,7 +469,7 @@ Utf8 Naming::niceArgumentRank(size_t idx)
     return form("argument [[%d]]", idx);
 }
 
-Utf8 Naming::niceParameterRank(size_t idx)
+Utf8 Naming::niceParameterRank(uint32_t idx)
 {
     switch (idx)
     {
